+++ Code/Examples/examples/src/main/scala/examples/demo/LFullyModularBall.scala
@@ -70,7 +70,7 @@ object LFullyModularBall extends Main {
     val inc = Clock.ticks.map(tick => velocity.readValueOnce * tick.toDouble)
 
     val pos = Events.foldAll(Pos(0,0))( acc => Events.Match(
-      reset >> { case Point(x, y) => Pos(x, y) },
+      reset >> { case Point(x, y) => Pos(x.toDouble, y.toDouble) },
       inc >>  { inc => acc + inc }
     ))
